Fix for non-unique country codes
The initial seeding somehow contains a duplicate country code, which is strange because I can't find any in the Geonames dataset. I did not explore the root cause any further.

The duplication causes the initial seeding to fail because of the unique constraint. With this fix the initial seeding is able to complete without further problems.